Create action1onpush.yml
On each push, runs community action codes named, that will The uses keyword tells the job to retrieve is v2 of the community action named actions/checkout@v2. This is an action that checks out your repository and downloads it to the runner, allowing you to run actions against your code (such as testing tools). You must use the checkout action any time your workflow will run against the repository's code or you are using an action defined in the repository.
    - uses: actions/setup-node@v1
This action installs the node software package on the runner, giving you access to the npm command.
    - run: npm install -g bats
The run keyword tells the job to execute a command on the runner. In this case, you are using npm to install the bats software testing package.
    - run: bats -v
Finally, you'll run the bats command with a parameter that outputs the software version.